escape separators in object keys and array values
In arrays: If the value is a string, escape the separators in the string so that we don't misread separators as encoding a serialized value when we go to deserialize.

In objects: Escape separators in the key even if the value isn't being serialized so that we don't misread separators as encoding a serialized value when we go to deserialize.