always encode type into value
Previously, we encoded the serializer type into object's keys when the value was in an object. This was different then what happens with values in arrays where the serializer type gets encoded directly into the value.

With this PR, we encode the serializer type into the value both in objects and arrays.